requirements.txt
Flask==2.2.2
Flask_Bcrypt==1.0.1
Flask_Mail==0.9.1
flask_sqlalchemy==3.0.3
itsdangerous==2.1.2
passlib==1.7.4
PyJWT==2.6.0
requests==2.28.1
SQLAlchemy==2.0.0
pillow==9.2.0
